IGNITE-18357 Move classes from fixtures to tests (#4122)
