+++ modules/raft/src/integrationTest/java/org/apache/ignite/raft/jraft/core/MockStateMachine.java
